Add "bazel4" overriding features
This updats the toolchain to build the arm64 sim under Bazel 4. The problem is
that Bazel should have a new CPU but that needs much work in Bazel. The idea of
this toolchain is to conditionally compile x86_64 simulator as arm64.

- bazel4.override_host_cpu_arm64
- bazel4.override_simulator_cpu_arm64

Why add a feature? There needs to be some additional context passed into Bazel
so it "knows" to replace the `x86_64`. These features could theoretically be
auto-configured on an ARM64 host but is a flag for the moment.